requirements.txt
evdev
pyaudio
nemo_toolkit["asr"]
cuda-python